requirements.txt
tensorflow==2.13.0
numpy
pandas
scipy
scikit-learn
matplotlib
fastapi
uvicorn
pydantic
google-cloud-storage
google-cloud-bigquery
google-cloud-aiplatform
google-cloud-monitoring
httpx
reportlab
pytest
